Extract methods from tablescreen for reuse in modules for redpotion.
