+++ lib/ProMotion/table/table_screen.rb
@@ -1,6 +1,7 @@
 module ProMotion
   class TableScreen < TableViewController
     include ProMotion::ScreenModule
+    include ProMotion::TableBuilder
     include ProMotion::Table
   end
 end
